LayoutTests/imported/w3c/web-platform-tests/css/css-scrollbars/scrollbar-width-keywords-expected.txt



FAIL scrollbar-width: none should have no scrollbar assert_equals: expected 100 but got 85
PASS scrollbar-width: thin should have scrollbar no wider than auto
